+++ src/main/java/com/hardware/entity/User.java
@@ -23,7 +23,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_user;
+    private Long id_user;
 
     @Column(nullable = false, unique = true)
     private String name;
